Move gift card details provider to providers dir
